split io_parser into submodule
