+++ app/processing/graph.py
@@ -4,7 +4,7 @@ Basic program graph used withing the :mod:`app.processing`.
 
 from __future__ import annotations
 
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from typing import TYPE_CHECKING
 from uuid import UUID, uuid4
