+++ app/processing/pre/io_parser/__init__.py
@@ -2,8 +2,6 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from io import UnsupportedOperation
-from itertools import chain
-from typing import Any, Generic, TypeVar, override
 
 from openqasm3.ast import (
     AliasStatement,
@@ -22,245 +20,11 @@ from openqasm3.ast import (
 from app.openqasm3.ast import QubitType
 from app.openqasm3.visitor import LeqoTransformer
 from app.processing.io_info import (
-    BitIOInfo,
     CombinedIOInfo,
-    QubitAnnotationInfo,
-    QubitIOInfo,
-    RegAnnotationInfo,
-    RegIOInfo,
-    RegSingleInputInfo,
-    RegSingleOutputInfo,
 )
-from app.processing.utils import expr_to_int, parse_io_annotation, parse_qasm_index
-
-T = TypeVar("T")
-RT = TypeVar("RT", bound=RegIOInfo[Any])
-
-
-class IOInfoBuilder(Generic[T], ABC):
-    """Abstract class for io info constructors for various types."""
-
-    @abstractmethod
-    def handle_declaration(
-        self,
-        declaration: QubitDeclaration | ClassicalDeclaration,
-        input_id: int | None,
-        dirty: bool = False,
-    ) -> None:
-        pass
-
-    @abstractmethod
-    def handle_alias(
-        self,
-        alias: AliasStatement,
-        output_id: int | None,
-        reusable: bool,
-    ) -> None:
-        pass
-
-    @abstractmethod
-    def finish(self) -> None:
-        pass
-
-
-class RegIOInfoBuilder(Generic[RT], IOInfoBuilder[RT], ABC):
-    next_id: int
-    io: RT
-    alias_to_ids: dict[str, list[int]]
-
-    def __init__(self, io: RT) -> None:
-        self.next_id = 0
-        self.io = io
-        self.alias_to_ids = {}
-
-    def identifier_to_ids(self, identifier: str) -> list[int] | None:
-        """Get ids via declaration_to_ids or alias_to_ids."""
-        result = self.io.declaration_to_ids.get(identifier)
-        return self.alias_to_ids.get(identifier) if result is None else result
-
-    def declaration_size_to_ids(self, size: Expression | None) -> list[int]:
-        reg_size = expr_to_int(size) if size is not None else 1
-        result = []
-        for _ in range(reg_size):
-            result.append(self.next_id)
-            self.next_id += 1
-        return result
-
-    def alias_expr_to_ids(
-        self,
-        value: Identifier | IndexExpression | Concatenation | Expression,
-    ) -> list[int] | None:
-        """Recursively get IDs list for alias expression."""
-        match value:
-            case IndexExpression():
-                collection = value.collection
-                if not isinstance(collection, Identifier):
-                    msg = f"Unsupported expression in alias: {type(collection)}"
-                    raise TypeError(msg)
-                source = self.identifier_to_ids(collection.name)
-                if source is None:
-                    return None
-                indices = parse_qasm_index([value.index], len(source))
-                return [source[i] for i in indices]
-            case Identifier():
-                return self.identifier_to_ids(value.name)
-            case Concatenation():
-                lhs = self.alias_expr_to_ids(value.lhs)
-                rhs = self.alias_expr_to_ids(value.rhs)
-                if lhs is None or rhs is None:
-                    return None
-                return lhs + rhs
-            case Expression():
-                msg = f"Unsupported expression in alias: {type(value)}"
-                raise UnsupportedOperation(msg)
-            case _:
-                msg = f"{type(value)} is not implemented as alias expression"
-                raise NotImplementedError(msg)
-
-
-class QubitIOInfoBuilder(RegIOInfoBuilder[QubitIOInfo]):
-    @override
-    def handle_declaration(
-        self,
-        declaration: QubitDeclaration | ClassicalDeclaration,
-        input_id: int | None,
-        dirty: bool = False,
-    ) -> None:
-        if not isinstance(declaration, QubitDeclaration):
-            msg = (
-                f"handle_declaration: expected QubitDeclaration got {type(declaration)}"
-            )
-            raise TypeError(msg)
-
-        name = declaration.qubit.name
-        qubit_ids = self.declaration_size_to_ids(declaration.size)
-
-        for i, qubit_id in enumerate(qubit_ids):
-            self.io.id_to_info[qubit_id] = QubitAnnotationInfo(
-                input=RegSingleInputInfo(input_id, i) if input_id is not None else None,
-                dirty=dirty,
-            )
-        self.io.declaration_to_ids[name] = qubit_ids
-
-        if input_id is not None:
-            self.io.input_to_ids[input_id] = qubit_ids
-        elif dirty:
-            self.io.dirty_ancillas.extend(qubit_ids)
-        else:  # non-input and non-dirty
-            self.io.required_ancillas.extend(qubit_ids)
-
-    @override
-    def handle_alias(
-        self,
-        alias: AliasStatement,
-        output_id: int | None,
-        reusable: bool,
-    ) -> None:
-        name = alias.target.name
-
-        qubit_ids = self.alias_expr_to_ids(alias.value)
-        if qubit_ids is None:
-            return
-        if len(qubit_ids) == 0:
-            msg = f"Unable to resolve IDs of alias {alias}"
-            raise RuntimeError(msg)
-
-        self.alias_to_ids[name] = qubit_ids
-        if output_id is not None:
-            self.io.output_to_ids[output_id] = qubit_ids
-
-        for i, qubit_id in enumerate(qubit_ids):
-            current_info = self.io.id_to_info[qubit_id]
-            if reusable:
-                if current_info.output is not None:
-                    msg = f"alias {name} declares output qubit as reusable"
-                    raise UnsupportedOperation(msg)
-                current_info.reusable = True
-                self.io.reusable_ancillas.append(qubit_id)
-            elif output_id is not None:
-                if current_info.output is not None:
-                    msg = f"alias {name} tries to overwrite already declared output"
-                    raise UnsupportedOperation(msg)
-                if current_info.reusable:
-                    msg = f"alias {name} declares output for reusable qubit"
-                    raise UnsupportedOperation(msg)
-                current_info.output = RegSingleOutputInfo(output_id, i)
-
-    @override
-    def finish(self) -> None:
-        returned_dirty = set(self.io.id_to_info.keys())
-        returned_dirty.difference_update(
-            self.io.reusable_ancillas,
-            self.io.reusable_after_uncompute,
-            set(chain(*self.io.output_to_ids.values())),
-        )
-        self.io.returned_dirty_ancillas = sorted(returned_dirty)
-
-
-class BitIOInfoBuilder(RegIOInfoBuilder[BitIOInfo]):
-    @override
-    def handle_declaration(
-        self,
-        declaration: QubitDeclaration | ClassicalDeclaration,
-        input_id: int | None,
-        dirty: bool = False,
-    ) -> None:
-        if not isinstance(declaration, ClassicalDeclaration) or not isinstance(
-            declaration.type,
-            BitType,
-        ):
-            msg = f"handle_declaration: expected bit declaration got {declaration}"
-            raise RuntimeError(msg)
-
-        name = declaration.identifier.name
-        bit_ids = self.declaration_size_to_ids(declaration.type.size)
-
-        for i, bit_id in enumerate(bit_ids):
-            self.io.id_to_info[bit_id] = RegAnnotationInfo(
-                input=RegSingleInputInfo(input_id, i) if input_id is not None else None,
-            )
-        self.io.declaration_to_ids[name] = bit_ids
-
-        if input_id is not None:
-            self.io.input_to_ids[input_id] = bit_ids
-
-    @override
-    def handle_alias(
-        self,
-        alias: AliasStatement,
-        output_id: int | None,
-        reusable: bool,
-    ) -> None:
-        name = alias.target.name
-
-        if reusable:
-            msg = (
-                f"Unsupported: reusable annotation over alias {name} referring to bits"
-            )
-            raise UnsupportedOperation(msg)
-
-        bit_ids = self.alias_expr_to_ids(alias.value)
-        if bit_ids is None:
-            return
-        if len(bit_ids) == 0:
-            msg = f"Unable to resolve IDs of alias {alias}"
-            raise RuntimeError(msg)
-
-        self.alias_to_ids[name] = bit_ids
-        if output_id is not None:
-            self.io.output_to_ids[output_id] = bit_ids
-
-        for i, bit_id in enumerate(bit_ids):
-            current_info = self.io.id_to_info[bit_id]
-            if output_id is not None:
-                if current_info.output is not None:
-                    msg = f"alias {name} tries to overwrite already declared output"
-                    raise UnsupportedOperation(msg)
-                current_info.output = RegSingleOutputInfo(output_id, i)
-
-    @override
-    def finish(self) -> None:
-        pass
+from app.processing.pre.io_parser.bits import BitIOInfoBuilder
+from app.processing.pre.io_parser.qubits import QubitIOInfoBuilder
+from app.processing.utils import parse_io_annotation
 
 
 class ParseAnnotationsVisitor(LeqoTransformer[None]):
@@ -451,4 +215,5 @@ class ParseAnnotationsVisitor(LeqoTransformer[None]):
         self.raise_on_non_contiguous_range(self.found_input_ids, "input")
         self.raise_on_non_contiguous_range(self.found_output_ids, "output")
         self.qubit_builder.finish()
+        self.bit_builder.finish()
         return node
